+++ src/api/middlewares/metrics_token.rs
