+++ src/api/mod.rs
@@ -1,2 +1,3 @@
 pub mod app;
+pub mod middlewares;
 pub mod routes;
