+++ src/ws/service/mod.rs
@@ -1,6 +1,6 @@
 pub mod database_handler;
 pub mod errors;
-pub mod friendship_event_handler;
+pub mod friendship_event_updates;
 pub mod friendship_event_validator;
 pub mod friendship_status_calculator;
 pub mod friendships_service;
